# REST APIs Design/openapi.yaml
openapi: 3.0.1
info:
  title: Swagger ToDo Manager
  description: REST APIs for the interaction with the ToDo Manager service.
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'https://todo.swagger.io/v2'
  - url: 'http://todo.swagger.io/v2'
...
paths:
...
  '/api/tasks/{taskId}/assignees':
    post:
      tags:
        - 'tasks/{taskId}/assignees'
      security:
        - cookieAuth: []
      summary: Assign a task to a user
      description: >-
        The task with ID taskId is assigned to the user specified in the request
        body. This operation can only be performed by the owner. For the user in the body email ad userId must be specified.
      operationId: assignTaskToUser
      parameters:
        - name: taskId
          in: path
          description: ID of the task
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: the user to whom the task is assigned
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '204':
          description: Successful operation - No Content
          content: {}
        '400':
          description: Bad Request
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    get:
      tags:
        - 'tasks/{taskId}/assignees'
      security:
        - cookieAuth: []
      summary: Retreve the users a task has been assigned to
      description: The users to whom the task with ID taskId is assigned are retrieved. This operation can be performed only by the owner.
      operationId: getUsersAssigned
      parameters:
        - name: taskId
          in: path
          description: ID of the task to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  '/api/tasks/{taskId}/assignees/{userId}':
    delete:
      tags:
        - 'tasks/{taskId}/assignees/{userId}'
      security:
        - cookieAuth: []
      summary: Remove a user from the assigned task
      description: >-
        The user that is removed is the user, identified by userId, that was
        assigned to the task identified by taskId. This operation can be performed only by the owner.
        If the user already completed the task, the operation is not performed because of a conflict.
        If the user is currently working on the task, the deletion is anyway performed.
      operationId: removeUser
      parameters:
        - name: taskId
          in: path
          description: ID of the assigned task
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          description: ID of the user to remove
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Successful operation - No Content
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '409':
          description: Conflict
          content: {}
        '500':
          description: Internal Server Error
          content: {}
...
  '/api/users/{userId}/selection':
      put:
        tags:
          - 'users/{userId}/selection'
        security:
          - cookieAuth: []
        summary: Select a task as the active task
        description: >-
          The task that is included in the request body is labeled as the active task for the user with ID userId, that must be currently logged in.
          The user must be an assignee of the task and must have not completed it yet.
          In the body, for the task, only id and description are compulsory.
          The selection of a task automatically remove previous selections.
        operationId: selectTask
        parameters:
          - name: userId
            in: path
            description: ID of the user that wants to select the task
            required: true
            schema:
              type: integer
              format: int64
        requestBody:
          description: the task to select as the active one
          content:
            application/json:
              schema:
               $ref: '#/components/schemas/Task'
        responses:
          '204':
            description: Successful operation - No Content
            content: {}
          '401':
            description: Unauthorized
            content: {}
          '403':
            description: Forbidden
            content: {}
          '404':
            description: Not Found
            content: {}
          '409':
            description: Conflict
            content: {}
          '500':
            description: Internal Server Error
            content: {}
  '/api/users/{userId}/deselection':
      put:
        tags:
          - 'users/{userId}/deselection'
        security:
          - cookieAuth: []
        summary: Deselect the currently active task
        description: >-
         The currently active task for the user with id userId is deselected. The userId must be currently logged in.
         If the user has no active task, a conflict error is raised.
        operationId: deselectTask
        parameters:
          - name: userId
            in: path
            description: ID of the user that wants to remove the selection of his/her own active task
            required: true
            schema:
              type: integer
              format: int64
        responses:
          '204':
            description: Successful operation - No Content
            content: {}
          '401':
            description: Unauthorized
            content: {}
          '403':
            description: Forbidden
            content: {}
          '409':
            description: Conflict
            content: {}
          '500':
            description: Internal Server Error
            content: {}
components:
  schemas:
    Task:
      type: object
      properties:
        $schema:
          type: string
          description: >-
            path to the JSON Schema to use for the validation of the JSON file
            describing the task
        id:
          type: integer
          description: unique identifier for the task
        description:
          type: string
          maxLength: 160
          description: textual task description
        important:
          type: boolean
          description: >-
            this property is set to true if the task is marked as important,
            false otherwise
          default: false
        private:
          type: boolean
          description: >-
            this property is set to true if the task is marked as private, false
            otherwise
          default: true
        project:
          type: string
          description: project in which the task is inserted
        deadline:
          type: string
          format: date-time
          description: the due date and hour of the task
        completed:
          type: boolean
          description: >-
            this property is set to true if the task is marked as completed,
            false otherwise
          default: false
        owner: 
          $ref: '#/components/schemas/User'
        assignees:
          type: array
          description: >-
            the list of users that has this task assigned
          items:
            $ref: '#/components/schemas/User'
        active:
          type: array
          description: >-
            the list of users that is currently working on this task
          items:
            $ref: '#/components/schemas/User'
        completer:
          type: array
          description: >-
            the list of users that have already completed this task
          items:
            $ref: '#/components/schemas/User'
        completers:
          type: integer
          description: >-
            minimum number of assegnee that must complete the task in order to
            consider it completed
          default: 1
        self:
          type: string
          format: uri-reference
          description: self link
    User:
      type: object
      properties:
        $schema:
          type: string
          description: >-
            path to the JSON Schema to use for the validation of the JSON file
            describing the task
        id:
          type: integer
          description: unique identifier for the user
        name:
          type: string
          description: username of the user
        email:
          type: string
          format: email
          description: email address of the user
        password:
          type: string
          minLength: 6
          maxLength: 20
          description: password of the user
        self:
          type: string
          format: uri-reference
          description: self link
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid